refactor components Configurator, Draw2DCanvas, file hooks.ts, created Canvas2DActions.tsx
